Move CharacterInterface to game/characters/base/interfaces
