+++ frontend/models/game/base/HitableInterface.php
@@ -9,6 +9,8 @@
 namespace frontend\models\game\base;
 
 
+use frontend\models\game\characters\base\interfaces\CharacterInterface;
+
 interface HitableInterface
 {
     public function beforeHit();
