+++ frontend/models/game/characters/Drone.php
@@ -10,7 +10,7 @@ namespace frontend\models\game\characters;
 
 
 use frontend\models\game\characters\base\Bee;
-use frontend\models\game\base\CharacterInterface;
+use frontend\models\game\characters\base\interfaces\CharacterInterface;
 use frontend\models\game\characters\interfaces\DroneInterface;
 
 class Drone extends Bee implements DroneInterface
