+++ frontend/models/game/characters/base/interfaces/CharacterInterface.php
@@ -6,8 +6,12 @@
  * Time: 03:33
  */
 
-namespace frontend\models\game\base;
+namespace frontend\models\game\characters\base\interfaces;
 
+
+use frontend\models\game\base\AliveInterface;
+use frontend\models\game\base\BeforeDeadInterface;
+use frontend\models\game\base\HitTakerInterface;
 
 interface CharacterInterface extends AliveInterface, HitTakerInterface, BeforeDeadInterface
 {
